# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Glazz API
  version: 0.2.0
  summary: HTTP contract for the Glazz MVP
  description: |
    Source of truth for Glazz HTTP resources. Chat generation is commanded over
    the versioned WebSocket protocol. Provider-specific concepts are private.
  contact:
    name: Glazz Engineering
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.glazz.hlab.sh/api/v1
    description: Production placeholder
...
security:
  - accessCookie: []
paths:
...
  /me:
    get:
      tags: [Authentication]
      summary: Get the current user
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user and effective permissions.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CurrentUser"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    patch:
      tags: [Authentication]
      summary: Update current-user preferences
      operationId: updateCurrentUser
      security: [{accessCookie: [], csrfHeader: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [locale]
              properties:
                locale: {type: string, enum: [en, es]}
      responses:
        "200":
          description: Updated current user.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CurrentUser"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    delete:
      tags: [Authentication]
      summary: Request account deletion
      operationId: requestAccountDeletion
      security: [{accessCookie: [], csrfHeader: [], recentAuth: []}]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirmation]
              properties:
                confirmation: {type: string, const: DELETE}
      responses:
        "202":
          description: Sessions revoked immediately; purge due within 24 hours.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DeletionJob"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "428":
          $ref: "#/components/responses/RecentAuthRequired"
...
  /me/deletion:
    get:
      tags: [Authentication]
      summary: Get account deletion state
      operationId: getAccountDeletion
      responses:
        "200":
          description: Current deletion job.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DeletionJob"}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /admin/models:
    get:
      tags: [Administration]
      summary: List models for administration
      operationId: adminListModels
      responses:
        "200":
          description: Internal catalog and exposure state; no secrets.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/AdminModel"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /admin/models/sync:
    post:
      tags: [Administration]
      summary: Synchronize provider model metadata
      operationId: adminSyncModels
      security: [{accessCookie: [], csrfHeader: []}]
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "202":
          description: Model synchronization enqueued; never auto-enables models.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Job"}
  /admin/models/{modelId}:
    patch:
      tags: [Administration]
      summary: Update model exposure
      operationId: adminUpdateModel
      security: [{accessCookie: [], csrfHeader: [], recentAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/ModelId"}
        - {$ref: "#/components/parameters/IfMatch"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateAdminModel"}
      responses:
        "200":
          description: Exposure changed and audit event committed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminModel"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "409": {$ref: "#/components/responses/Conflict"}
        "428": {$ref: "#/components/responses/RecentAuthRequired"}
...
  /admin/settings/{key}:
    patch:
      tags: [Administration]
      summary: Update a runtime setting
      operationId: adminUpdateSetting
      security: [{accessCookie: [], csrfHeader: [], recentAuth: []}]
      parameters:
        - {name: key, in: path, required: true, schema: {type: string, pattern: "^[a-z][a-z0-9_.-]+$"}}
        - {$ref: "#/components/parameters/IfMatch"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [value]
              properties:
                value: {}
      responses:
        "200":
          description: Setting updated and audited.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RuntimeSetting"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "409": {$ref: "#/components/responses/Conflict"}
        "428": {$ref: "#/components/responses/RecentAuthRequired"}
  /admin/users:
    get:
      tags: [Administration]
      summary: List users for administration
      operationId: adminListUsers
      parameters:
        - {$ref: "#/components/parameters/Limit"}
        - {$ref: "#/components/parameters/After"}
        - {name: query, in: query, schema: {type: string, maxLength: 200}}
      responses:
        "200":
          description: User metadata without conversation content.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/AdminUser"}
                  nextCursor: {type: [string, "null"]}
  /admin/users/{userId}/role:
    patch:
      tags: [Administration]
      summary: Update a user role
      operationId: adminUpdateUserRole
      security: [{accessCookie: [], csrfHeader: [], recentAuth: []}]
      parameters:
        - {name: userId, in: path, required: true, schema: {$ref: "#/components/schemas/Id"}}
        - {$ref: "#/components/parameters/IfMatch"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role: {$ref: "#/components/schemas/Role"}
      responses:
        "200":
          description: Role changed without allowing last-admin removal.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminUser"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "409": {$ref: "#/components/responses/Conflict"}
        "428": {$ref: "#/components/responses/RecentAuthRequired"}
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: glazz_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: glazz_refresh
    guestCookie:
      type: apiKey
      in: cookie
      name: glazz_guest
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    recentAuth:
      type: apiKey
      in: cookie
      name: glazz_recent_auth
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Quoted current resource version.
      schema: {type: string, pattern: '^"[1-9][0-9]*"$'}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 30}
    After:
      name: after
      in: query
      schema: {type: string, minLength: 1}
...
    ModelId:
      name: modelId
      in: path
      required: true
      schema: {$ref: "#/components/schemas/Id"}
  responses:
    BadRequest:
      description: Invalid request.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
    Unauthorized:
      description: Missing, expired, or invalid actor credentials.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
    Forbidden:
      description: Actor lacks permission.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
    NotFound:
      description: Resource does not exist or is not owned by the actor.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
    Conflict:
      description: Resource state, idempotency, or optimistic version conflict.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
...
    RecentAuthRequired:
      description: Recent Google reauthentication is required.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
...
  schemas:
    Id:
      type: string
      pattern: "^[0-9a-fA-F-]{36}$"
...
    Model:
      type: object
      required: [id, name, capabilities]
      properties:
        id: {$ref: "#/components/schemas/Id"}
        name: {type: string}
        description: {type: string}
        capabilities:
          type: object
          required: [chatCompletions, markdown, code]
          properties:
            chatCompletions: {type: boolean, const: true}
            markdown: {type: boolean}
            code: {type: boolean}
...
    Role: {type: string, enum: [user, admin]}
    CurrentUser:
      type: object
      required: [id, email, displayName, locale, role, plan, permissions]
      properties:
        id: {$ref: "#/components/schemas/Id"}
        email: {type: string, format: email}
        displayName: {type: string}
        avatarUrl: {type: [string, "null"], format: uri}
        locale: {type: string, enum: [en, es]}
        role: {$ref: "#/components/schemas/Role"}
        plan: {type: string, enum: [free, pro]}
        permissions: {type: array, items: {type: string}, uniqueItems: true}
...
    DeletionJob:
      type: object
      required: [id, status, requestedAt, dueAt]
      properties:
        id: {$ref: "#/components/schemas/Id"}
        status: {type: string, enum: [pending, processing, completed, failed]}
        requestedAt: {type: string, format: date-time}
        dueAt: {type: string, format: date-time}
        completedAt: {type: [string, "null"], format: date-time}
    Job:
      type: object
      required: [id, status]
      properties:
        id: {$ref: "#/components/schemas/Id"}
        status: {type: string, enum: [pending, processing, completed, failed]}
    AdminModel:
      allOf:
        - {$ref: "#/components/schemas/Model"}
        - type: object
          required: [enabled, available, supported, audience, defaultFor, order, version]
          properties:
            enabled: {type: boolean}
            available: {type: boolean}
            supported: {type: boolean}
            audience: {type: array, items: {type: string, enum: [guest, user]}}
            defaultFor: {type: array, items: {type: string, enum: [guest, user]}}
            order: {type: integer, minimum: 0}
            version: {type: integer, minimum: 1}
    UpdateAdminModel:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        enabled: {type: boolean}
        audience: {type: array, items: {type: string, enum: [guest, user]}, uniqueItems: true}
        defaultFor: {type: array, items: {type: string, enum: [guest, user]}, uniqueItems: true}
        order: {type: integer, minimum: 0}
    RuntimeSetting:
      type: object
      required: [key, value, version, updatedAt]
      properties:
        key: {type: string}
        value:
          oneOf:
            - {type: boolean}
            - {type: integer}
            - {type: string}
            - {type: array, items: {type: string}}
        version: {type: integer, minimum: 1}
        updatedAt: {type: string, format: date-time}
    AdminUser:
      type: object
      required: [id, email, displayName, role, status, version, createdAt]
      properties:
        id: {$ref: "#/components/schemas/Id"}
        email: {type: string, format: email}
        displayName: {type: string}
        role: {$ref: "#/components/schemas/Role"}
        status: {type: string, enum: [active, deletion_pending, disabled]}
        version: {type: integer, minimum: 1}
        createdAt: {type: string, format: date-time}
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - unauthenticated
                - forbidden
                - not_found
                - conflict
                - quota_exceeded
                - concurrency_exceeded
                - global_budget_exceeded
                - maintenance
                - dependency_unavailable
                - precondition_required
                - recent_auth_required
                - refresh_token_reused
            message: {type: string}
            requestId: {type: string}
            details: {type: object, additionalProperties: true}